requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
idna==3.6
jsonlines==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
